[refactor] 방명록 좋아요 로직 위치 변경
